Update README.md
In this project, named "newsletter-sign", I decided to test the use of Tailwind CSS in conjunction with JavaScript and React. The goal was to create a subscription screen for a newsletter. When an invalid email is entered, an error message is displayed. On the other hand, when a valid email is entered, the user is redirected to a confirmation screen.